+++ lib/src/androidTest/java/com/github/thibaultbee/srtdroid/models/Socket6Test.kt
@@ -37,7 +37,6 @@ class Socket6Test {
 
     @Before
     fun setUp() {
-        assert(Srt.startUp() >= 0)
         socket = Socket()
         assertTrue(socket.isValid)
     }