+++ lib/src/androidTest/java/com/github/thibaultbee/srtdroid/models/transmission/SimpleStreamTest.kt
@@ -40,7 +40,6 @@ class SimpleStreamTest {
 
     @Before
     fun setUp() {
-        assert(Srt.startUp() >= 0)
         socket = Socket()
         assertTrue(socket.isValid)
     }