+++ lib/src/main/java/com/github/thibaultbee/srtdroid/enums/RejectReasonCode.kt
@@ -15,10 +15,12 @@
  */
 package com.github.thibaultbee.srtdroid.enums
 
+import com.github.thibaultbee.srtdroid.Srt
 import com.github.thibaultbee.srtdroid.models.rejectreason.InternalRejectReason
 
 /**
  * Internal reject reason for [InternalRejectReason].
+ * Once it has been called, you must release Srt context with [Srt.cleanUp] when application leaves.
  *
  * **See Also:** [Reject Reasons](https://github.com/Haivision/srt/blob/master/docs/API-functions.md#rejection-reasons-1)
  */
@@ -127,5 +129,9 @@ enum class RejectReasonCode {
          * User defined error codes offset
          */
         const val USERDEFINED_OFFSET = 2000 //  SRT_REJC_USERDEFINED
+
+        init {
+            Srt.startUp()
+        }
     }
 }